+++ frontend/lib/seo/content-momentum.ts
@@ -6,14 +6,32 @@ import type {
   SeoSourceMetrics,
   StrategicSeoFamily,
 } from './internal-seo-types';
-import { classifyMissingContentIntent } from './missing-content';
 import {
   detectStrategicModelFamily,
-  getBusinessPriorityWeight,
   getSeoFamilyStatus,
-  normalizeSeoQuery,
 } from './seo-intents';
 import { clusterGscQueries, stripOrigin, summarizeRows } from './seo-opportunity-engine';
+import {
+  formatContentMomentumMarkdown,
+  formatDelta,
+  formatSignedNumber,
+  formatSignedPercent,
+} from './content-momentum-format';
+import {
+  calibrateScore,
+  classifyMixedFamilyMomentum,
+  classifyTrend,
+  isJunkMomentumDraft,
+  isWorthMomentumItem,
+  scoreMomentumDraft,
+  scoreToPriority,
+} from './content-momentum-scoring';
+
+export {
+  formatContentMomentumMarkdown,
+  formatContentMomentumSectionMarkdown,
+  labelizeMomentumType,
+} from './content-momentum-format';
 
 const MAX_MOMENTUM_ITEMS = 80;
 const STRATEGIC_FAMILIES = new Set(['Seedance', 'Kling', 'Veo', 'LTX']);
@@ -23,7 +41,7 @@ type BuildContentMomentumOptions = {
   previousRows: GscPerformanceRow[];
 };
 
-type MomentumDraft = {
+export type MomentumDraft = {
   type: ContentMomentumType;
   pageUrl: string | null;
   queryCluster: string | null;
@@ -52,62 +70,6 @@ export function buildContentMomentumItems(options: BuildContentMomentumOptions):
   )
     .sort((a, b) => priorityRank(a.priority) - priorityRank(b.priority) || b.score - a.score)
     .slice(0, MAX_MOMENTUM_ITEMS);
-}
-
-export function formatContentMomentumMarkdown(item: ContentMomentumItem): string {
-  return [
-    'Title:',
-    buildMomentumTitle(item),
-    '',
-    'Target:',
-    item.pageUrl ? stripOrigin(item.pageUrl) : item.queryCluster ?? item.family,
-    '',
-    'Source:',
-    item.queryCluster ? `GSC query cluster: "${item.queryCluster}"` : `GSC family/page momentum: ${item.family}`,
-    `Current: ${formatMetrics(item.current)}`,
-    `Previous: ${formatMetrics(item.previous)}`,
-    `Delta: ${formatDelta(item.clickDelta, 'clicks')}, ${formatDelta(item.impressionDelta, 'impressions')}, ${formatSignedPercent(item.ctrDelta)} CTR, ${formatSignedNumber(item.positionDelta)} position`,
-    '',
-    'Observed trend:',
-    item.observedTrend,
-    '',
-    'Recommended action:',
-    item.recommendedAction,
-    '',
-    'Why it matters:',
-    item.whyItMatters,
-    '',
-    'Acceptance criteria:',
-    ...item.acceptanceCriteria.map((criterion) => `- ${criterion}`),
-  ].join('\n');
-}
-
-export function formatContentMomentumSectionMarkdown(items: ContentMomentumItem[]): string {
-  if (!items.length) {
-    return ['# Content Momentum', '', 'No Content Momentum items generated for this snapshot.'].join('\n');
-  }
-
-  return [
-    '# Content Momentum',
-    '',
-    `Generated items: ${items.length}`,
-    '',
-    ...items.map((item, index) => [`## ${index + 1}. ${buildMomentumTitle(item)}`, '', formatContentMomentumMarkdown(item)].join('\n')),
-  ].join('\n\n');
-}
-
-export function labelizeMomentumType(value: ContentMomentumType) {
-  if (value === 'gaining_page') return 'Gaining page';
-  if (value === 'declining_page') return 'Declining page';
-  if (value === 'gaining_cluster') return 'Gaining cluster';
-  if (value === 'declining_cluster') return 'Declining cluster';
-  if (value === 'rising_family') return 'Rising family';
-  if (value === 'declining_family') return 'Declining family';
-  if (value === 'mixed_family_momentum') return 'Mixed family momentum';
-  if (value === 'refresh_candidate') return 'Refresh candidate';
-  if (value === 'protect_winner') return 'Protect winner';
-  if (value === 'outdated_model_attention') return 'Older model attention';
-  return 'Watchlist';
 }
 
 function buildPageMomentumDrafts(currentRows: GscPerformanceRow[], previousRows: GscPerformanceRow[]): MomentumDraft[] {
@@ -291,117 +253,6 @@ function finalizeMomentumItem(draft: MomentumDraft): ContentMomentumItem {
   };
 }
 
-function scoreMomentumDraft(
-  draft: MomentumDraft,
-  clickDelta: number,
-  impressionDelta: number,
-  ctrDelta: number
-) {
-  const maxImpressions = Math.max(draft.current.impressions, draft.previous.impressions);
-  const deltaMagnitude = Math.abs(impressionDelta) + Math.abs(clickDelta) * 8;
-  const volumeScore = Math.min(34, Math.log10(Math.max(maxImpressions, 1)) * 14);
-  const deltaScore = Math.min(34, Math.log10(Math.max(deltaMagnitude, 1)) * 16);
-  const positionScore = draft.current.averagePosition <= 6 ? 14 : draft.current.averagePosition <= 12 ? 10 : draft.current.averagePosition <= 30 ? 3 : -10;
-  const ctrScore = draft.type.includes('declining') || draft.type === 'refresh_candidate' ? Math.min(10, Math.max(0, -ctrDelta * 220)) : Math.min(8, Math.max(0, ctrDelta * 180));
-  return Math.round((volumeScore + deltaScore + positionScore + ctrScore + draft.scoreBoost) * getBusinessPriorityWeight(draft.family));
-}
-
-function calibrateScore(draft: MomentumDraft, rawScore: number) {
-  const status = getSeoFamilyStatus(draft.family);
-  const maxImpressions = Math.max(draft.current.impressions, draft.previous.impressions);
-  const impressionDelta = Math.abs(draft.current.impressions - draft.previous.impressions);
-  const currentImpressions = draft.current.impressions;
-  const clickDelta = Math.abs(draft.current.clicks - draft.previous.clicks);
-  if (draft.family === 'Brand') {
-    return Math.min(rawScore, isBrandTypoMomentum(draft) ? 75 : 103);
-  }
-  if (maxImpressions < 35 || impressionDelta < 20) return Math.min(rawScore, 48);
-  if (draft.type === 'mixed_family_momentum') return Math.min(rawScore, 48);
-  if ((draft.type === 'declining_cluster' || draft.type === 'declining_page') && currentImpressions < 20 && maxImpressions < 80) {
-    return Math.min(rawScore, 48);
-  }
-  if (currentImpressions < 50 && maxImpressions < 90) return Math.min(rawScore, 75);
-  if (draft.current.averagePosition >= 30 && maxImpressions < 250) return Math.min(rawScore, 48);
-  if (draft.type === 'watchlist') return Math.min(rawScore, 48);
-  if (draft.type === 'outdated_model_attention') return Math.min(rawScore, 48);
-  if (draft.family === 'Sora') return Math.min(rawScore, 48);
-  if (status === 'emerging' && maxImpressions < 120) return Math.min(rawScore, 48);
-  if (status === 'emerging') return Math.min(rawScore, 62);
-  if ((draft.type === 'rising_family' || draft.type === 'declining_family') && maxImpressions < 100) return Math.min(rawScore, 58);
-  if (rawScore >= 104 && !isCriticalEligible(draft, impressionDelta, clickDelta)) return Math.min(rawScore, 103);
-  return rawScore;
-}
-
-function scoreToPriority(score: number): SeoActionPriority {
-  if (score >= 104) return 'critical';
-  if (score >= 76) return 'high';
-  if (score >= 52) return 'medium';
-  return 'low';
-}
-
-function classifyTrend(current: SeoSourceMetrics, previous: SeoSourceMetrics): 'gaining' | 'declining' | null {
-  const maxImpressions = Math.max(current.impressions, previous.impressions);
-  const impressionDelta = current.impressions - previous.impressions;
-  const clickDelta = current.clicks - previous.clicks;
-  const relativeImpressionDelta = previous.impressions ? impressionDelta / previous.impressions : current.impressions ? 1 : 0;
-  const relativeClickDelta = previous.clicks ? clickDelta / previous.clicks : current.clicks ? 1 : 0;
-
-  if (maxImpressions < 20) return null;
-  if (Math.abs(impressionDelta) < 20 && Math.abs(clickDelta) < 4) return null;
-
-  if (
-    (impressionDelta >= 40 && relativeImpressionDelta >= 0.25) ||
-    (clickDelta >= 5 && relativeClickDelta >= 0.25)
-  ) {
-    return 'gaining';
-  }
-  if (
-    (impressionDelta <= -40 && relativeImpressionDelta <= -0.25) ||
-    (clickDelta <= -5 && relativeClickDelta <= -0.25)
-  ) {
-    return 'declining';
-  }
-  return null;
-}
-
-function classifyMixedFamilyMomentum(current: SeoSourceMetrics, previous: SeoSourceMetrics): boolean {
-  const impressionDelta = current.impressions - previous.impressions;
-  const clickDelta = current.clicks - previous.clicks;
-  const positionDelta = current.averagePosition - previous.averagePosition;
-  const clicksRising = clickDelta >= 5;
-  const impressionsFalling = impressionDelta <= -40;
-  const positionWorse = positionDelta >= 3 && current.averagePosition > 12;
-  const clicksFalling = clickDelta <= -5;
-  const impressionsRising = impressionDelta >= 40;
-  return (clicksRising && (impressionsFalling || positionWorse)) || (clicksFalling && impressionsRising);
-}
-
-function isWorthMomentumItem(draft: MomentumDraft) {
-  const maxImpressions = Math.max(draft.current.impressions, draft.previous.impressions);
-  const impressionDelta = Math.abs(draft.current.impressions - draft.previous.impressions);
-  const clickDelta = Math.abs(draft.current.clicks - draft.previous.clicks);
-  const status = getSeoFamilyStatus(draft.family);
-
-  if (draft.type === 'outdated_model_attention') return draft.current.impressions >= 50;
-  if (maxImpressions < 20) return false;
-  if (impressionDelta < 20 && clickDelta < 4) return false;
-  if (status === 'emerging') return maxImpressions >= 35 && impressionDelta >= 15;
-  if (draft.type === 'mixed_family_momentum') return maxImpressions >= 80 && (impressionDelta >= 40 || clickDelta >= 5);
-  if (draft.type === 'rising_family' || draft.type === 'declining_family') return maxImpressions >= 80 && impressionDelta >= 50;
-  return true;
-}
-
-function isJunkMomentumDraft(draft: MomentumDraft) {
-  const text = normalizeSeoQuery([
-    draft.pageUrl,
-    draft.queryCluster,
-    draft.family,
-    ...draft.representativeQueries,
-  ].filter(Boolean).join(' '));
-  if (isAdultOrJunkText(text)) return true;
-  return draft.representativeQueries.some((query) => classifyMissingContentIntent(query) === 'irrelevant_junk');
-}
-
 function buildObservedTrend(
   draft: MomentumDraft,
   clickDelta: number,
@@ -485,16 +336,6 @@ function buildAcceptanceCriteria(draft: MomentumDraft) {
     criteria.push('Underlying query/page clusters are reviewed before assigning implementation work');
   }
   return criteria;
-}
-
-function buildMomentumTitle(item: ContentMomentumItem) {
-  if (item.type === 'protect_winner') return `Protect growing SEO winner: ${item.pageUrl ? stripOrigin(item.pageUrl) : item.queryCluster}`;
-  if (item.type === 'refresh_candidate') return `Refresh declining SEO page: ${item.pageUrl ? stripOrigin(item.pageUrl) : item.queryCluster}`;
-  if (item.type === 'outdated_model_attention') return `Reposition older ${item.family} demand`;
-  if (item.type === 'rising_family') return `${item.family} family is gaining momentum`;
-  if (item.type === 'declining_family') return `${item.family} family is losing momentum`;
-  if (item.type === 'mixed_family_momentum') return `${item.family} family has mixed momentum`;
-  return `${labelizeMomentumType(item.type)}: ${item.queryCluster ?? item.pageUrl ?? item.family}`;
 }
 
 function groupRowsByPage(rows: GscPerformanceRow[]) {
@@ -599,49 +440,8 @@ function priorityRank(priority: SeoActionPriority) {
   return 4;
 }
 
-function isCriticalEligible(draft: MomentumDraft, impressionDelta: number, clickDelta: number) {
-  const hasAbsoluteSignal = draft.current.impressions >= 250 || impressionDelta >= 150;
-  const hasClickSignal = draft.current.clicks >= 20 || clickDelta >= 15;
-  const nearPageOne = draft.current.averagePosition > 0 && draft.current.averagePosition <= 12;
-  const strategic = STRATEGIC_FAMILIES.has(draft.family);
-  if (hasAbsoluteSignal && hasClickSignal && nearPageOne) return true;
-  return strategic && nearPageOne && draft.current.impressions >= 120 && impressionDelta >= 80 && hasClickSignal;
-}
-
-function isBrandTypoMomentum(draft: MomentumDraft) {
-  const text = normalizeSeoQuery([draft.queryCluster, ...draft.representativeQueries].filter(Boolean).join(' '));
-  return /\bmaxvedio\b|\bmaxvideos\b|\bmax videoa\b|\bmaxvideai\b|\bmax vidio\b/.test(text);
-}
-
 function targetLabel(draft: MomentumDraft) {
   return draft.queryCluster ?? draft.pageUrl ?? draft.family;
-}
-
-function formatMetrics(metrics: SeoSourceMetrics) {
-  return [
-    `${metrics.clicks} clicks`,
-    `${metrics.impressions} impressions`,
-    `${(metrics.ctr * 100).toFixed(2)}% CTR`,
-    `avg position ${metrics.averagePosition.toFixed(1)}`,
-  ].join(', ');
-}
-
-function formatDelta(value: number, label: string) {
-  return `${formatSignedNumber(value)} ${label}`;
-}
-
-function formatSignedNumber(value: number) {
-  if (!Number.isFinite(value)) return '0';
-  return `${value > 0 ? '+' : ''}${value.toFixed(Number.isInteger(value) ? 0 : 1)}`;
-}
-
-function formatSignedPercent(value: number) {
-  if (!Number.isFinite(value)) return '+0.00%';
-  return `${value > 0 ? '+' : ''}${(value * 100).toFixed(2)}%`;
-}
-
-function isAdultOrJunkText(value: string) {
-  return /\b(?:porn|porno|nsfw|nude|naked|sex|sexy|xxx|onlyfans|casino|gambling|torrent|crack|hack)\b/.test(value);
 }
 
 function stableId(value: string) {
