+++ frontend/lib/seo/unified-action-briefs.ts
@@ -16,6 +16,9 @@ import type {
 } from './internal-seo-types';
 import { compactIntentLabel } from './seo-intents';
 import { stripOrigin } from './seo-opportunity-engine';
+import { formatUnifiedActionBriefMarkdown, labelizeIssue } from './unified-action-brief-format';
+
+export { formatUnifiedActionBriefMarkdown, formatUnifiedActionBriefsMarkdown } from './unified-action-brief-format';
 
 type BuildUnifiedActionBriefsOptions = {
   opportunities?: StrategicSeoOpportunity[];
@@ -213,45 +216,6 @@ export function buildUnifiedActionBriefs(options: BuildUnifiedActionBriefsOption
     .map(finalizeBrief)
     .filter((brief) => shouldKeepBrief(brief))
     .sort((a, b) => priorityWeight[b.priority] - priorityWeight[a.priority] || b.score - a.score || a.targetUrl.localeCompare(b.targetUrl));
-}
-
-export function formatUnifiedActionBriefMarkdown(brief: UnifiedPageActionBrief): string {
-  return [
-    'Title:',
-    titleForBrief(brief),
-    '',
-    'Target:',
-    brief.targetUrl,
-    '',
-    'Source signals:',
-    ...brief.combinedSignals.map((signal) => `- ${labelizeModule(signal.module)}: ${signal.summary}`),
-    '',
-    'Problem:',
-    brief.observedProblem,
-    '',
-    'Recommended implementation:',
-    brief.recommendedImplementation,
-    '',
-    'Supporting actions:',
-    ...brief.supportingActions.map((action) => `- ${action}`),
-    '',
-    'Risks / caveats:',
-    ...brief.risks.map((risk) => `- ${risk}`),
-    '',
-    'Acceptance criteria:',
-    ...brief.acceptanceCriteria.map((criterion) => `- ${criterion}`),
-  ].join('\n');
-}
-
-export function formatUnifiedActionBriefsMarkdown(briefs: UnifiedPageActionBrief[]): string {
-  if (!briefs.length) return ['# Recommended Page Actions', '', 'No unified page action briefs generated for this snapshot.'].join('\n');
-  return [
-    '# Recommended Page Actions',
-    '',
-    `Generated briefs: ${briefs.length}`,
-    '',
-    ...briefs.map((brief, index) => [`## ${index + 1}. ${titleForBrief(brief)}`, '', formatUnifiedActionBriefMarkdown(brief)].join('\n')),
-  ].join('\n\n');
 }
 
 function getDraft(
@@ -496,26 +460,4 @@ function briefKey(targetUrl: string, queryCluster: string, family: string, inten
 
 function stableId(parts: string[]) {
   return `brief-${parts.join('|').toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/^-|-$/g, '').slice(0, 96)}`;
-}
-
-function titleForBrief(brief: Pick<UnifiedPageActionBrief, 'queryCluster' | 'targetUrl' | 'intent'>) {
-  if (brief.intent === 'prompt_examples' || brief.intent === 'prompt_guide' || brief.intent === 'examples') {
-    return `Strengthen ${brief.targetUrl} for ${brief.queryCluster}`;
-  }
-  if (brief.intent === 'comparison') return `Expand/defend ${brief.queryCluster} on ${brief.targetUrl}`;
-  if (brief.intent === 'brand_typo') return `Defend brand typo demand on ${brief.targetUrl}`;
-  return `Improve ${brief.targetUrl} for ${brief.queryCluster}`;
-}
-
-function labelizeModule(module: UnifiedActionSourceModule) {
-  if (module === 'opportunity_finder') return 'Opportunity Finder';
-  if (module === 'ctr_doctor') return 'CTR Doctor';
-  if (module === 'missing_content') return 'Missing Content';
-  if (module === 'internal_links') return 'Internal Links';
-  if (module === 'url_inspection') return 'URL Inspection';
-  return 'Momentum';
-}
-
-function labelizeIssue(value: string) {
-  return value.replace(/_/g, ' ').replace(/\b\w/g, (letter) => letter.toUpperCase());
 }
